+++ resources/js/routes.js
@@ -1,10 +1,13 @@
-import Passwords from './views/Passwords'
+
 import NotFound from './views/errors/NotFound'
 import Dashboard from './views/Dashboard'
-import Roles from './views/Roles'
-import Role from "./views/Role";
-import Users from "./views/Users";
-import User from "./views/User";
+
+import ProfilePasswords from './views/Profile/Passwords'
+
+import AdminRolesIndex from './views/Admin/Roles/Index'
+import AdminRolesShow from "./views/Admin/Roles/Show";
+import AdminUsersIndex from "./views/Admin/Users/Index";
+import AdminUsersShow from "./views/Admin/Users/Show";
 
 export default [
     {
@@ -20,7 +23,7 @@ export default [
     {
         path: '/settings/password',
         name: 'settings.password',
-        component: Passwords,
+        component: ProfilePasswords,
     },
 
 
@@ -32,7 +35,7 @@ export default [
     {
         path: '/admin/roles',
         name: 'admin.roles',
-        component: Roles,
+        component: AdminRolesIndex,
         meta: {
             permission: 'auth.roles'
         }
@@ -40,7 +43,7 @@ export default [
     {
         path: '/admin/role/new',
         name: 'admin.role.create',
-        component: Role,
+        component: AdminRolesShow,
         meta: {
             permission: 'auth.roles.create'
         }
@@ -48,7 +51,7 @@ export default [
     {
         path: '/admin/role/:id',
         name: 'admin.role.update',
-        component: Role,
+        component: AdminRolesShow,
         meta: {
             permission: 'auth.roles.update'
         }
@@ -58,7 +61,7 @@ export default [
     {
         path: '/admin/users',
         name: 'admin.users',
-        component: Users,
+        component: AdminUsersIndex,
         meta: {
             permission: 'auth.users'
         }
@@ -67,7 +70,7 @@ export default [
     {
         path: '/admin/users/:id',
         name: 'admin.users.view',
-        component: User,
+        component: AdminUsersShow,
         meta: {
             permission: 'auth.users'
         }
@@ -83,7 +86,12 @@ export default [
 
 // Don't want the globally applied loading spinner on a route?
 // add in the following meta field:
-
 // meta: {
 //     globalLoadState: false
+// }
+
+// Route Permissions
+// These can be added by specifying the permission meta field
+// meta: {
+//     permission: 'auth.users'
 // }
