+++ resources/js/views/Admin/Roles/Index.vue
@@ -31,8 +31,8 @@
 </template>
 <script>
     import gql from 'graphql-tag'
-    import Errors from '../components/ui/errors'
-    import DefaultLayout from "../components/layout/DefaultLayout";
+    import Errors from '../../../components/ui/errors'
+    import DefaultLayout from "../../../components/layout/DefaultLayout";
 
     export default {
         components: {DefaultLayout},
